bookly_backend_1dec/requirements.txt
alembic==1.17.2
attr==0.3.2
ConfigParser==7.2.0
cryptography==46.0.3
debugpy==1.8.17
dnspython==2.8.0
docutils==0.22.3
fakeredis==2.32.1
fastapi==0.123.0
fastapi_mail==1.5.8
filelock==3.20.0
Flask==3.1.2
HTMLParser==0.0.2
ipython==9.7.0
ipywidgets==8.1.8
itsdangerous==2.2.0

keyring==25.7.0
passlib==1.7.4
protobuf==6.33.1
pycparser==2.23
pydantic==2.12.5
pydantic_settings==2.12.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
smtp_email_sender==1.0.4
Sphinx==9.0.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.50.0
thread==2.0.5
typing_extensions==4.15.0
urllib3_secure_extra==0.1.0
watchfiles==1.1.1

